net: ParkMap builds under WebAssembly and passes CI's analyser
- ParkMap::home shifted the hash's top bits down by a member that started at
  64 for an empty table. No lookup runs before the first growth sets it, but
  clang-analyzer-core.BitwiseShift cannot see that and CI's clang-tidy
  refused it. The product's upper half is now masked to the table: a
  constant shift, and no dependence on std::size_t's width.
- ParkTable_test.cpp narrowed a std::uint64_t index to std::size_t
  implicitly, which is 32 bits under Emscripten: -Wshorten-64-to-32.

+++ src/core/net/detail/ParkTable_test.cpp
@@ -96,7 +96,8 @@ TEST_CASE("ParkMap agrees with an unordered_map over a long random run", "[net][
         }
         else if (choice == 1)
         {
-            auto const index = rng() % live.size();
+            // Narrowed explicitly: `std::size_t` is 32 bits under WebAssembly.
+            auto const index = static_cast<std::size_t>(rng() % live.size());
             auto const id = live[index];
             auto const taken = map.erase(ParkId { id });
             REQUIRE(taken.get() == reference.at(id));
